Feature/homeview portions calculation

+++ Snusless/ViewModels/EditProfileViewModel.swift
@@ -14,12 +14,12 @@ class EditProfileViewModel {
     var editName: String = ""
     var originalName: String = ""
     var editStartDate: Date = Date()
-    var editDosor: Int = 0
+    var editDosor: Double = 0.0
     var editPrice: Double = 0.0
     var editPortioner: Int = 0
     var editSavingsGoal: Int = 0
     
-    func updateUser(context: ModelContext, updatedName: String, updatedDate: Date, updatedDosor: Int, updatedPortions: Int, updatedPrice: Double, updatedSavingsGoal: Int) {
+    func updateUser(context: ModelContext, updatedName: String, updatedDate: Date, updatedDosor: Double, updatedPortions: Int, updatedPrice: Double, updatedSavingsGoal: Int) {
             do {
                 let descriptor = FetchDescriptor<User>()
                 if let existingUser = try context.fetch(descriptor).first {
